INC-38 refactor: split type file into comments and posts
